NONE: Decompose `JiraService` and expose additional service-level errors
