+++ src/infrastructure/jira/errors.ts
@@ -1,0 +1,5 @@
+import { CauseAwareError } from '../../common/errors';
+
+export class ForbiddenByJiraServiceError extends CauseAwareError {}
+
+export class NotFoundInJiraServiceError extends CauseAwareError {}
