+++ src/infrastructure/jira/index.ts
@@ -1,1 +1,4 @@
+export { ConfigurationState } from './jira-app-configuration-service';
+export { JiraSubmitDesignServiceError } from './jira-design-service';
 export * from './jira-service';
+export * from './errors';
