+++ src/infrastructure/jira/jira-design-service.ts
@@ -1,0 +1,139 @@
+import type { SubmitDesignsResponse } from './jira-client';
+import { jiraClient } from './jira-client';
+
+import { CauseAwareError } from '../../common/errors';
+import { truncate } from '../../common/string-utils';
+import type {
+	AtlassianDesign,
+	ConnectInstallation,
+} from '../../domain/entities';
+import { AtlassianAssociation } from '../../domain/entities';
+
+export type SubmitDesignParams = {
+	readonly design: AtlassianDesign;
+	readonly addAssociations?: AtlassianAssociation[];
+	readonly removeAssociations?: AtlassianAssociation[];
+};
+
+export class JiraDesignService {
+	/**
+	 * @throws {JiraSubmitDesignServiceError} Design submission fails.
+	 */
+	submitDesign = async (
+		params: SubmitDesignParams,
+		connectInstallation: ConnectInstallation,
+	): Promise<void> => {
+		return this.submitDesigns([params], connectInstallation);
+	};
+
+	/**
+	 * @throws {JiraSubmitDesignServiceError} Design submission fails.
+	 */
+	submitDesigns = async (
+		designs: SubmitDesignParams[],
+		connectInstallation: ConnectInstallation,
+	): Promise<void> => {
+		const associationsLastUpdated = new Date();
+
+		// Jira does not allow display names longer than 255 characters.
+		const MAX_DISPLAY_NAME_LENGTH = 255;
+		const response = await jiraClient.submitDesigns(
+			{
+				designs: designs.map(
+					({ design, addAssociations, removeAssociations }) => ({
+						...design,
+						displayName: truncate(design.displayName, MAX_DISPLAY_NAME_LENGTH),
+						addAssociations: addAssociations ?? null,
+						removeAssociations: removeAssociations ?? null,
+						associationsLastUpdated: associationsLastUpdated.toISOString(),
+						associationsUpdateSequenceNumber: associationsLastUpdated.valueOf(),
+					}),
+				),
+			},
+			connectInstallation,
+		);
+
+		this.throwIfSubmitDesignResponseHasErrors(response);
+	};
+
+	/**
+	 * @throws {JiraSubmitDesignServiceError}
+	 */
+	private throwIfSubmitDesignResponseHasErrors = (
+		response: SubmitDesignsResponse,
+	) => {
+		if (response.rejectedEntities.length) {
+			const { key, errors } = response.rejectedEntities[0];
+			throw JiraSubmitDesignServiceError.designRejected(key.designId, errors);
+		}
+
+		// TODO: Confirm whether we need to consider the use case below as a failure and throw or just leave a warning.
+		if (response.unknownIssueKeys?.length) {
+			throw JiraSubmitDesignServiceError.unknownIssueKeys(
+				response.unknownIssueKeys,
+			);
+		}
+
+		if (response.unknownAssociations?.length) {
+			throw JiraSubmitDesignServiceError.unknownAssociations(
+				response.unknownAssociations.map(
+					(x) => new AtlassianAssociation(x.associationType, x.values),
+				),
+			);
+		}
+	};
+}
+
+export const jiraDesignService = new JiraDesignService();
+
+export class JiraSubmitDesignServiceError extends CauseAwareError {
+	designId?: string;
+	rejectionErrors?: { readonly message: string }[];
+	unknownIssueKeys?: string[];
+	unknownAssociations?: AtlassianAssociation[];
+
+	private constructor({
+		message,
+		designId,
+		rejectionErrors,
+		unknownIssueKeys,
+		unknownAssociations,
+	}: {
+		message: string;
+		designId?: string;
+		rejectionErrors?: { readonly message: string }[];
+		unknownIssueKeys?: string[];
+		unknownAssociations?: AtlassianAssociation[];
+	}) {
+		super(message);
+		this.designId = designId;
+		this.rejectionErrors = rejectionErrors;
+		this.unknownIssueKeys = unknownIssueKeys;
+		this.unknownAssociations = unknownAssociations;
+	}
+
+	static designRejected(
+		designId: string,
+		rejectionErrors: { readonly message: string }[],
+	): JiraSubmitDesignServiceError {
+		return new JiraSubmitDesignServiceError({
+			message: 'The design submission has been rejected',
+			designId,
+			rejectionErrors,
+		});
+	}
+
+	static unknownIssueKeys(unknownIssueKeys: string[]) {
+		return new JiraSubmitDesignServiceError({
+			message: 'The design has unknown issue keys',
+			unknownIssueKeys,
+		});
+	}
+
+	static unknownAssociations(unknownAssociations: AtlassianAssociation[]) {
+		return new JiraSubmitDesignServiceError({
+			message: 'The design has unknown associations',
+			unknownAssociations,
+		});
+	}
+}
